fix: fix makefiles
Clean some whitespaces.

Remove -e for echo because it is non-standard (bmake).

Add 'all' dependency for main Makefile's install so it is possible to
directly do `make install` instead of `make all install`.

Fix clean so it doesn't delete gencodecs objects and fix purge so it
'clean' and call 'clean' for gencodecs as I assume it was how it
supposed to work.