Fix permanent shutdowns (TLS & 1000 close code) (#168)
* fix(gateway): not reconnecting on 1000 close codes

This commit fixes the issue that it wouldn't reconnect on 1000 close codes.

* fix(websockets): not calling on_close on TLS notify

This commit fixes the issue that once a TLS notify was received, it wouldn't call on_close, ending up stuck in loop, not reconnecting.

* fix(loop): not breaking loop on `DISCORD_SESSION_SHUTDOWN`

This commit fixes an issue that the discord-loop loop wouldn't be breaken to reconnect, causing the bot to be stuck in the state of disconnected.

* chore(websocket): update `already closed` ws message

This commit updated the message that is printed to the terminal when a `ws_close` is called on a connection that has already ended.

* revert(gateway): unnecessary cleanup

This commit reverts a cleanup that is unnecessary as it's expected that `discord_cleanup` is called after `discord_run`.

---------

Co-authored-by: Lucas Müller <lucas@muller.codes>